Changed main to be able to point to routes, and file restructuring
